refactor: StockSpecPage hook 커스텀으로 분리
